requirements.txt
# ==========================================================
#  TextureForge – AI Texture Upscaler
#  Requirements
# ==========================================================

# Core AI / ML
torch>=2.0.0
torchvision>=0.15.0
basicsr>=1.4.2
realesrgan>=0.3.0
facexlib>=0.3.0
gfpgan>=1.3.8

# Image processing
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0

# UI
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0
PyQt6-sip>=13.5.0

# torch.compile acceleration (Windows)
triton-windows>=3.0.0; sys_platform == 'win32'

# Optional format support
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
